feat: Add elevation preservation to projected points
Adds the option to keep the original elevation of structural
points when projecting them onto the section in qProf_QWidget.py and
modifies the `map_struct_pts_on_section` function in
gis_utils/intersections.py to support this functionality.
Also includes a fix for the `SourceLine2DLayerDialog` to refresh
attribute and feature selection comboboxes.  Adds dem elevation to
qProf_plotting